Lift out mvi verfication
