+++ modules/claims_api/app/controllers/claims_api/concerns/mvi_verification.rb
@@ -1,0 +1,19 @@
+
+# frozen_string_literal: true
+
+module ClaimsApi
+  module MviVerification
+    extend ActiveSupport::Concern
+
+    included do
+      before_action :verify_mvi
+
+      def verify_mvi
+        unless target_veteran.mvi_record?
+          render json: { errors: [{ detail: 'Not found' }] },
+                 status: :not_found
+        end
+      end
+    end
+  end
+end
